chatbot/requirements.txt
# EV Chatbot - Requirements
# Python 3.9+

# AI/ML
google-generativeai==0.3.2

# Web Interface
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
